refactor(admin): extract ProductsGridRowView to satisfy max-lines guard
products-grid.tsx crossed the 500-line threshold after the dynamic
column rewrite and pushed the frozen baseline (21) to 22. The row view
(structural cells + model-driven data cell dispatch) moves to
products-grid-row.tsx; no behaviour change.

Refs #2387
